Cores refactor + file condensation
