+++ examples/research/hdbscan_validation_paper.py
@@ -28,11 +28,11 @@ from tqdm import tqdm
 import nomad.data as data_folder
 import nomad.io.base as loader
 import nomad.stop_detection.utils as utils
-import nomad.stop_detection.dbscan as TADBSCAN
-import nomad.stop_detection.hdbscan as HDBSCAN
-import nomad.stop_detection.grid_based as GRID_BASED
-import nomad.stop_detection.lachesis as LACHESIS
-import nomad.stop_detection.density_based as SEQSCAN
+import nomad.stop_detection.density_algs as TADBSCAN
+import nomad.stop_detection.density_algs as HDBSCAN
+import nomad.stop_detection.sequential_algs as GRID_BASED
+import nomad.stop_detection.sequential_algs as LACHESIS
+import nomad.stop_detection.density_algs as SEQSCAN
 from nomad.stop_detection.validation import (
     AlgorithmRegistry,
     bootstrap_metric_summary,
